implementations/python/pyproject.toml
[project]
name = "md-osi-fiscal"
version = "1.0.0rc1"
description = "Implementación de referencia Python de md-osi-fiscal-mx (Open Semantic Interchange Fiscal MX)."
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
readme = "README.md"
authors = [
    { name = "MD Consultoría SC", email = "mdsamca2025@gmail.com" }
]
dependencies = [
    "pyyaml>=6",
    "jsonschema>=4",
    "pydantic>=2",
]

[project.urls]
Homepage = "https://github.com/MarxCha/md-osi-fiscal-mx"
Repository = "https://github.com/MarxCha/md-osi-fiscal-mx"
Issues = "https://github.com/MarxCha/md-osi-fiscal-mx/issues"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["md_osi_fiscal*"]
